fix error

+++ models.py
@@ -59,7 +59,7 @@ class BaseJetGraphGCN(LightningModule):
     def validation_step(self, batch, batch_idx):
         out = self(batch)  # Perform a single forward pass.
         predictions = torch.sigmoid(out)
-        labels = batch.y.unsqueeze(1)
+        labels = batch.y.unsqueeze(1).float()
 
         print('labels and pred shapes:', labels.shape, predictions.shape)
